refactor: reestructure singleton editor folders and names
