+++ src/components/Campaign.test.tsx
@@ -9,7 +9,7 @@ import { Campaign, editorTopBarTestId, errorMessageTestId } from "./Campaign";
 import { TestDopplerIntlProvider } from "./i18n/TestDopplerIntlProvider";
 import { Design } from "react-email-editor";
 import { act, screen, waitFor } from "@testing-library/react";
-import { SingletonDesignContext } from "./SingletonEditor";
+import { SingletonDesignContextProvider } from "./singleton-editor/singletonDesignContext";
 import userEvent from "@testing-library/user-event";
 import { Result } from "../abstractions/common/result-types";
 import { CampaignContent } from "../abstractions/domain/content";
@@ -145,13 +145,13 @@ const createTestContext = () => {
         }}
       >
         <TestDopplerIntlProvider>
-          <SingletonDesignContext.Provider value={singletonEditorContext}>
+          <SingletonDesignContextProvider value={singletonEditorContext}>
             <MemoryRouter initialEntries={[routerInitialEntry]}>
               <Routes>
                 <Route path="/:idCampaign" element={<Campaign />} />
               </Routes>
             </MemoryRouter>
-          </SingletonDesignContext.Provider>
+          </SingletonDesignContextProvider>
         </TestDopplerIntlProvider>
       </AppServicesProvider>
     </QueryClientProvider>
