+++ src/components/singleton-editor/index.ts
@@ -1,0 +1,2 @@
+export { SingletonEditorProvider } from "./SingletonEditorProvider";
+export { useSingletonEditor } from "./useSingletonEditor";
